+++ TOMB4/game/draw.h
@@ -32,7 +32,7 @@ void GetRoomBounds();
 void SetRoomBounds(short* door, int rn, ROOM_INFO* actualRoom);
 void DrawEffect(short fx_num);
 void PrintObjects(short room_number);
-int GetFrames(ITEM_INFO* item, short* frm[], long* rate);
+int GetFrames(ITEM_INFO* item, short* frm[], int* rate);
 void GetBoundsAccurate(ITEM_INFO* item, short* result);
 short* GetBestFrame(ITEM_INFO* item);
 void UpdateSkyLightning();
