+++ TOMB4/game/rope.h
@@ -11,7 +11,7 @@ typedef struct COLL_INFO COLL_INFO;
 void DrawRopeList();
 void ProjectRopePoints(ROPE_STRUCT* Rope);
 PHD_VECTOR* Normalise(PHD_VECTOR* v);
-void phd_GetMatrixAngles(long* m, short* dest);
+void phd_GetMatrixAngles(int* m, short* dest);
 void AlignLaraToRope(ITEM_INFO* l);
 void SetPendulumVelocity(int x, int y, int z);
 void RopeControl(short item_num);
